% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_TPP.R
\name{analyse_TPP}
\alias{analyse_TPP}
\title{Analyse TPP-TR Data}
\usage{
analyse_TPP(
  TPP_tbl,
  comparisons = NULL,
  quantity_column = "rel_quantity",
  control_name = "Control",
  silent = FALSE,
  ...
)
}
\arguments{
\item{TPP_tbl}{A data frame (or \link{tibble}) containing proteomics data from a
thermal protein profiling (TPP) experiment. This must contain the columns:
\itemize{
\item \code{Protein_ID}: Unique protein identity
\item \code{Condition} : Category of treated or control sample
\item \code{Replicate} : Replicate number.
\item \code{Temperature}: Experiment temperature in \eqn{\degree C}
\item Protein quantity measurement corresponding to the non-denatured fraction
}

If replicates are not given, a single replicate per condition will be
assumed.}

\item{comparisons}{A data.frame containing the information needed to build
the comparisons to use to calculate \eqn{\Delta T_m} and other statistics.
Columns must be \emph{Condition_01}, \emph{Replicate_01},
\emph{Condition_02} and \emph{Replicate_01}, \emph{e.g}:\tabular{rrrr}{
   Condition_01 \tab Replicate_01 \tab Condition_02 \tab Replicate_01 \cr
   Treated \tab 01 \tab Control \tab 01 \cr
   Treated \tab 02 \tab Control \tab 02 \cr
}


If no table is given, one will be generated based on the conditions in the
data, comparing each condition to control and like-for-like replicate
comparisons.}

\item{quantity_column}{Character. Name for the column containing
protein quantity values in \code{TPP_tbl}}

\item{control_name}{Character. Character string that matches the control
experiment in \code{TPP_tbl$Condition}}

\item{silent}{Boolean. If \code{TRUE}, no console output is shown}

\item{...}{Further arguments to be passed to \code{\link[=fit_melt_by_experiment]{fit_melt_by_experiment()}} and
\code{\link[=nls_multstart]{nls_multstart()}}}
}
\value{
A \code{tibble} containing all data in TPP_tbl, with additional calculated
columns (where possible) detaling curve parameters and comparison statistics
}
\description{
\code{analyse_TPP()} transforms \emph{Thermal Protein Profiling} (TPP) relative
intensity data, fitting sigmoidal melting curves to the presumably
normalised data, generating curve parameters - including melting points,
then calculating statistical parameters including p-value as described by
Savitsky \emph{et al.} 2014:
\itemize{
\item Melting point differences (\eqn{\Delta T_m}) are found and filtered to
those for which \eqn{R^2 > 0.8} for all observations and
\eqn{plateau < 0.3} for all control curves.
\item Proteins are ordered by steepness of melting curve slope - steepest first.
\item These proteins are then divided into bins of 300, and the final group
added to the second-to-last if less than 300.
\item Per bin, the left- and right-sided robust standard deviation is estimated
using the 15.87, 50, and 84.13 percentiles and calculating p-values for
all measurements binwise.
\item The p-values thus calculated are adjusted by applying the
Benjamini-Hochberg procedure to control the false discovery rate
(\emph{FDR}).
}
}
\examples{
# Minimal data - two-protein melt curve
x <- quan_data_normP
norm_x <- normalise_TPP(x)

# Analyse - single core curve fitting
analyse_TPP(norm_x, max_cores = 1)

# Custom comparisons, e.g. compare treated replicates
comparison_tbl <-  data.frame(
  Condition_01 = c("Treated", "Treated", "Treated"),
  Replicate_01 = c("01", "02", "01"),
  Condition_02 = c("Control", "Control","Treated") ,
  Replicate_02 = c("01", "02", "02")
)

analyse_TPP(
  norm_x,
  comparisons = comparison_tbl,
  max_cores = 1
)
}
\references{
Savitski M. M. \emph{et al.}, Tracking cancer drugs in living cells by
thermal profiling of the proteome. \emph{Science}, 346: 1255784 (2014)

Benjamini, Y., and Hochberg, Y. Controlling the false discovery
rate: a practical and powerful approach to multiple testing. \emph{Journal
of the Royal Statistical Society Series B}, 57, 289–300 (1995)
}
