% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{analysis_data_4prot}
\alias{analysis_data_4prot}
\title{Statistics for 4 Protein Melting Curve}
\format{
A data frame with 16 rows and 19 variables:
\describe{
\item{Protein_ID}{Gene name identifying the protein}
\item{Condition}{Treatment given to this sample set}
\item{Replicate}{Replicate number for this sample set}
\item{F_scaled}{Scaled F-statistic from NPARC analysis}
\item{p_adj_NPARC}{p-value from NPARC analysis, FDR adjusted using
\emph{Benjamini-Hochberg} correction}
\item{\code{a}}{Parameter \code{a} from sigmoidal curve fitting}
\item{\code{b}}{Parameter \code{b} from sigmoidal curve fitting}
\item{melt_point}{Melting point, calculated as the model temperature
resulting in 50 \% of the protein destabilised}
\item{infl_point}{Inflection point, calculated as the model temperature
for which \eqn{f''(T_{infl}) = 0 }}
\item{slope}{Sigmoidal curve slope, calculated as \eqn{f'(T_{infl}) }}
\item{plateau}{The plateau of the modeled sigmoidal curve}
\item{R_sq}{\eqn{R^2} for the fitted sigmoidal curve}
\item{Comparison}{Measurements compared for \eqn{\Delta T_m}}
\item{diff_melt_point}{Melting point difference (\eqn{\Delta T_m})}
\item{min_slope}{Minimum slope across protein}
\item{min_comparison_slope}{Minimum slope for comparison of non-control
\emph{vs} control}
\item{min_R_sq}{Minimum \eqn{R^2} across protein}
\item{max_control_plateau}{Maximum curve plateau in control samples for
each protein}
\item{adj_pvalue}{p-value for melting point difference, FDR adjusted using
\emph{Benjamini-Hochberg} correction}}
}
\usage{
analysis_data_4prot
}
\description{
A dataset containing the statistics and parameters generated from fitting a
melting curve to normalised proteomics data (DIA) for the 4 proteins and
comparing melting curves across measurements:
}
\keyword{datasets}
