% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_observed_TPP_plot.R
\name{build_observed_TPP_plot}
\alias{build_observed_TPP_plot}
\title{Build Plot from TPP observed Protein data}
\usage{
build_observed_TPP_plot(
  data,
  facets = TRUE,
  quan_column = "rel_quantity",
  colour_column = "Condition",
  shape_column = "Replicate",
  facets_column = "Protein_ID"
)
}
\arguments{
\item{data}{A data frame (or \link{tibble}) containing proteomics data from a
thermal protein profiling (TPP) experiment. In all cases, this must contain
the columns:
\describe{
\item{Temp}{Experiment temperature in \eqn{\degree C}, will be used for the x
axis values}
\item{rel_quantity}{(or name specified with \code{quan_column}) Protein quantity
measurement corresponding to \verb{Fraction non-denatured}, will be used for the
y-axis values}
}
The following columns will be used for the \code{ggplot} \code{aesthetics}, but
default values will be assumed if not present:
\describe{
\item{Protein_ID}{Gene name identifying the protein - used for facets and
to provide values for default \code{Condition} column; defaults to generic
"Protein"}
\item{Condition}{Treatment given to this sample set - used for colour;
defaults to Protein_ID values}
\item{Replicate}{Replicate number for this sample set - used for shape;
defaults to "01"}
}}

\item{facets}{Boolean. Whether to separate facets using \emph{Protein_ID}}

\item{quan_column}{Character. The column in \code{data} and \code{predicted_data} used
for plot y-values}

\item{colour_column}{Character. The column in \code{data} and \code{predicted_data} used
for plot colour aesthetic}

\item{shape_column}{Character. The column in \code{data} and \code{predicted_data} used
for plot shape aesthetic}

\item{facets_column}{Character. The column in \code{data} and \code{predicted_data} used
for plot facets}
}
\value{
A \code{ggplot} object.
}
\description{
\code{build_observed_TPP_plot()} creates a \link{ggplot2} plot, showing the protein
melting data points given in \code{data} only. To add fitted melting curves use
\code{build_melt_plot_with_model()}.
}
\examples{
# Minimal data - four-protein melt curve (Protein A only)
x <- quan_data_4prot[quan_data_4prot$Protein_ID == "Protein_A",]

# Create plot using default appearance
build_observed_TPP_plot(x)

# Specify column names
build_observed_TPP_plot(
  data = x,
  quan_column = "rel_quantity",
  colour_column = "Condition",
  shape_column = "Replicate",
  facets_column = "Protein_ID"
  )

# Remove facets
build_observed_TPP_plot(
  data = x,
  facets = FALSE
  )

# As a ggplot object, additions can be made to returned plot
# For example, changing the theme:
build_observed_TPP_plot(x) +
  ggplot2::theme_minimal()
}
