% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_TPP.R
\name{export_TPP}
\alias{export_TPP}
\title{Export full TPP results data}
\usage{
export_TPP(
  TPP_data,
  file_name,
  path = NULL,
  format = c("xlsx", "csv", "tsv", "Rda"),
  include_columns = NULL,
  silent = FALSE
)
}
\arguments{
\item{TPP_data}{Data to export in a \code{data.frame} or \code{tibble}}

\item{file_name}{Character. File name to be written to. If \code{format} is not
set, file extension will be used to decide format.}

\item{path}{Character. Optionally, the directory path to the location to be
saved in, if not the working directory.}

\item{format}{File save format:
\itemize{
\item \emph{xlsx:} Excel spreadsheet, using \code{\link[writexl:write_xlsx]{writexl::write_xlsx()}}
\item \emph{csv:} Comma-separated text file
\item \emph{tsv:} Tab-separated text file
\item \emph{Rda:} Rdata saved object - will loaded into R session as
"TPP_results"
}

If not given, will default to "xlsx"}

\item{include_columns}{Character vector. Column names to included in the
exported data. If not given, the all columns in \code{TPP_data} will be exported}

\item{silent}{Boolean. If true, will run silently, without console output.}
}
\value{
\code{TPP_data}, in the format supplied
}
\description{
\code{export_TPP()} exports a results data frame in the chosen format.
}
\examples{
# Minimal data - analysed two-protein melt curve
x <- MP_stat_data_ATIC

# Default behaviour - save a .xlsx spreadsheet of results data.
file <- tempfile()
export_TPP(x, file)
unlink(paste0(file, ".xlsx"))

# Passing a filename with a .tsv extension saves as tab-separated data
file <- tempfile(fileext = ".tsv")
export_TPP(x, file)
unlink(file)

# Explicitly save as .csv text file
file <- tempfile()
export_TPP(
  x,
  file,
  format = "csv"
)
unlink(paste0(file, ".csv"))

# Save Rdata object with no console messages
file <- tempfile()
export_TPP(
  x,
  file,
  format = "Rda",
  silent = TRUE
  )
unlink(paste0(file, ".Rda"))
}
