% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_melt_by_experiment.R
\name{fit_melt_by_experiment}
\alias{fit_melt_by_experiment}
\title{Fit Melting Curve for Experimental Groups}
\usage{
fit_melt_by_experiment(
  data,
  experiment_cols = c("Protein_ID", "Condition", "Replicate"),
  silent = FALSE,
  max_cores = 8,
  ...
)
}
\arguments{
\item{data}{A data frame (or \link{tibble}) containing proteomics data from a
thermal protein profiling (TPP) experiment. This must contain the columns:
\itemize{
\item Experiment temperature in \eqn{\degree C}
\item Protein quantity measurement corresponding to the non-denatured fraction
}}

\item{experiment_cols}{A character vector of column names, which will be used
to split the data into constituent experimental groups. This should include
a column that uniquely identifies proteins in the data}

\item{silent}{Boolean. If \code{TRUE}, no console output is shown}

\item{max_cores}{Integer. The maximum number of cores to parallelise the
curve fitting over, if less than the maximum number of cores available.
If \code{max_cores} is 1, then curve fitting will be run serially in
a single process}

\item{...}{Further arguments to be passed to \code{\link[=fit_melting_curve]{fit_melting_curve()}} and
\code{\link[=nls_multstart]{nls_multstart()}}}
}
\value{
A \code{tibble} with 1 row and \emph{13 + n} variables (where
\emph{n} is the length of \code{experiment_cols}):
\itemize{
\item ID columns from \code{experiment_cols}
\item \code{a}: Parameter \code{a} from fitted sigmoidal curve.
\item \code{b}: Parameter \code{b} from fitted sigmoidal curve.
\item \code{plateau}: Plateau from fitted sigmoidal curve.
\item \code{melt_point}:  Melting point (\eqn{T_m}), temperature at which 50 \% of the
protein is destabilised
\item \code{infl_point}: Inflection point (\eqn{T_{infl}}), temperature at which the
second derivative \eqn{f''(T) = 0}
\item \code{slope}: Sigmoidal curve slope, \emph{i.e.} \eqn{f'(T_{infl})}
\item \code{R_sq}: \eqn{R^2} for the fitted sigmoidal curve
\item \code{RSS}: Residual sum of squares (\emph{RSS})
\item \code{sigma}: Residual standard deviation (\eqn{\sigma})
\item \code{n_coeffs}: Number of coefficients fitted
\item \code{n_obs}: Number of observations in curve
\item \code{log_lik}: log-likelihood (\eqn{\ell})
\item \code{AICc}: Corrected Akaike information criterion (\emph{AICc})
}
}
\description{
\code{fit_melt_by_experiment()} fits sigmoidal melting curves to all the proteins
in \code{data} using \code{\link[=fit_melting_curve]{fit_melting_curve()}}, splitting into individual curves
identified with supplied columns.

Both serial and parallel processing (using \code{PSOCK}) of the individual curves
are supported.
}
\examples{
# Minimal data - four-protein melt curve
x <- quan_data_4prot

# Fit melting curves - 2-core processing
fit_melt_by_experiment(x, max_cores = 2)

# Single-core (serial) processing
fit_melt_by_experiment(x, max_cores = 1)

# Specify experiment details to separate.
fit_melt_by_experiment(
  x,
  experiment_cols = c("Protein_ID", "Condition", "Replicate"),
  max_cores = 2
)

}
