% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_NPARC_pval.R
\name{get_NPARC_pval}
\alias{get_NPARC_pval}
\title{Get p-value by nonparametric analysis of response curves \emph{(NPARC)}}
\usage{
get_NPARC_pval(
  TPP_tbl,
  degrees_of_freedom = c(4:7),
  max_cores = 4,
  comparisons = NULL,
  quantity_column = "quantity",
  control_name = "Control",
  to_plot = FALSE,
  to_save = NULL,
  silent = FALSE
)
}
\arguments{
\item{TPP_tbl}{A data frame (or \link{tibble}) containing proteomics data from a
thermal protein profiling (TPP) experiment. This must contain the columns:
\itemize{
\item \code{Protein_ID}: Unique protein identity
\item \code{Condition} : Category of treated or control sample
\item \code{Replicate} : Replicate number.
\item \code{Temperature}: Experiment temperature in \eqn{\degree C}
\item Protein quantity measurement corresponding to the non-denatured fraction
}

If replicates are not given, a single replicate per condition will be
assumed.}

\item{degrees_of_freedom}{Numeric. Range of degrees of freedom to use to
generate possible \eqn{H_{null}} and \eqn{H_{alt}} curves.}

\item{max_cores}{Integer. The maximum number of cores to parallelise the
spline fitting over, if less than the maximum number of cores available.
If \code{max_cores} is 1, then spline fitting will be run serially in
a single process}

\item{comparisons}{A data.frame containing the information needed to build
the comparisons to use to calculate \eqn{\Delta T_m} and other statistics.
Columns must be \emph{Condition_01}, \emph{Replicate_01},
\emph{Condition_02} and \emph{Replicate_01}, \emph{e.g}:\tabular{rrrr}{
   Condition_01 \tab Replicate_01 \tab Condition_02 \tab Replicate_01 \cr
   Treated \tab 01 \tab Control \tab 01 \cr
   Treated \tab 02 \tab Control \tab 02 \cr
}


If no table is given, one will be generated based on the conditions in the
data, comparing each condition to control and like-for-like replicate
comparisons.}

\item{quantity_column}{Character. Name for the column containing
protein quantity values in \code{TPP_tbl}}

\item{control_name}{Character. Character string that matches the control
experiment in \code{TPP_tbl$Condition}}

\item{to_plot}{Boolean. If true, will generate distribution plots of
F-statistic and p-value on calculating \emph{NPARC} statistics}

\item{to_save}{Character. If supplied, will generate distribution plots of
F-statistic and p-value and save with \code{\link[=ggsave]{ggsave()}}}

\item{silent}{Boolean. If \code{TRUE}, no console output is shown}
}
\value{
A \code{tibble}, with the columns \code{Protein_ID}, \code{Condition}, \code{F-scaled},
and \code{p_adj_NPARC}
}
\description{
Calculate F-Score and p-value with nonparametric analysis of response curves
\emph{(NPARC)} as described by Childs \emph{et al.} 2019:
\itemize{
\item For each protein and condition comparison, assume a null hypothesis
\eqn{H_{null}} (measurements from both conditions can be modelled with the
same sigmoidal melting curve) and alternate hypothesis \eqn{H_{alt}}
(measurements each condition can be modelled with separate sigmoidal
melting curve.)
\item Curves are fitted to each hypothesis for each protein, and the degrees
of freedom estimated for each curve fit.
\item The F-statistic and p-value is computed using the residual sum of squares
\emph{RSS} from \eqn{H_{alt}} and \eqn{H_{null}} calculated degrees
of freedom for for each protein and condition comparison.
\item p-values are adjusted by applying the
Benjamini-Hochberg procedure to control the false discovery rate
(\emph{FDR}).
}
}
\examples{
# Minimal data - two-protein melt curve
x <- quan_data_normP
norm_x <- normalise_TPP(x)

# NPARC F-score and p-value generation
get_NPARC_pval(norm_x,
               quantity_column = "rel_quantity",
               max_cores = 1)

# Plot F-score and p-value distribution
get_NPARC_pval(norm_x,
               quantity_column = "rel_quantity",
               max_cores = 1,
               to_plot = TRUE)
}
\references{
Childs, D., \emph{et al.} Non-Parametric Analysis of Thermal Proteome
Profiles Reveals Novel Drug-Binding Proteins. \emph{Molecular & Cellular
Proteomics}, 18, 2506-2515, (2019)

Benjamini, Y., and Hochberg, Y. Controlling the false discovery
rate: a practical and powerful approach to multiple testing. \emph{Journal
of the Royal Statistical Society Series B}, 57, 289–300 (1995)
}
