% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pval_by_melting_point.R
\name{get_pval_by_melting_point}
\alias{get_pval_by_melting_point}
\title{Get p-value by comparison of modelled melting point differences}
\usage{
get_pval_by_melting_point(fit_tbl)
}
\arguments{
\item{fit_tbl}{A data frame (or \code{tibble}) of parameters from fitting melting
curves to each protein and taking melting point differences. This must
include the columns:
\itemize{
\item \code{Protein_ID} - Unique protein identifier
\item \code{Condition} - Experimental conditio \emph{e.g.} "Control", "Treatment"
\item \code{min_R_sq} - Minimum coefficient of determination (\eqn{R^2}) for all
observations of a protein
\item \code{max_control_plateau} - Maximum plateau parameter (\eqn{pl}) across all
control curves for a protein
\item \code{diff_melt_point} - The difference in melting point (\eqn{\Delta T_m})
between two observations of a protein, \emph{e.g.} between a treated
and control sample
\item \code{min_slope} - the slope (\eqn{f'(T_{infl})}) of the steepest curve per
comparison
}}
}
\value{
A \code{tibble} containing all data in fit_tbl, with additional calculated
column for FDR-adjusted p-value (\code{adj_pvalue})
}
\description{
Calculate p-value from melting-point differences as described by Savitsky
\emph{et al.} 2014:
\itemize{
\item Melting point differences (\eqn{\Delta T_m}) are found and filtered to
those for which \eqn{R^2 > 0.8} for all observations and
\eqn{plateau < 0.3} for all control curves.
\item Proteins are ordered by steepness of melting curve slope - steepest first.
\item These proteins are then divided into bins of 300, and the final group
added to the second-to-last if less than 300.
\item Per bin, the left- and right-sided robust standard deviation is estimated
using the 15.87, 50, and 84.13 percentiles and calculating p-values for
all measurements binwise.
\item The p-values thus calculated are adjusted by applying the
Benjamini-Hochberg procedure to control the false discovery rate
(\emph{FDR}).
}
}
\examples{
# Minimal data - ATIC metling curve statistics
x <- MP_stat_data_ATIC

# Melting point p-value generation
get_pval_by_melting_point(x)

}
\references{
Savitski M. M. \emph{et al.} Tracking cancer drugs in living cells by
thermal profiling of the proteome. \emph{Science}, 346: 1255784 (2014)

Benjamini, Y., and Hochberg, Y. Controlling the false discovery
rate: a practical and powerful approach to multiple testing. \emph{Journal
of the Royal Statistical Society Series B}, 57, 289-300 (1995)
}
