% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_TPP_hits.R
\name{get_TPP_hits}
\alias{get_TPP_hits}
\title{Get Hit Protein Results from TPP-TR data}
\usage{
get_TPP_hits(
  TPP_data,
  n_hits = NULL,
  hit_criteria = "default_hit_criteria",
  control_name = "Control",
  to_export = "TPP_hits.xlsx",
  to_plot = FALSE,
  to_save = NULL,
  plot_separately = FALSE,
  silent = FALSE,
  ...
)
}
\arguments{
\item{TPP_data}{TPP-TR Analysis results data from \code{analyse_TPP}. Must have
\code{melt_point}, \code{diff_melt_point}, and \code{adj_pvalue} columns.

\code{min_comparison_slope} (the steepest slope parameter between the plots
compared for each \eqn{\Delta T_m} comparison) is necessary for the
\code{slope_threshold} hit criteria to be in effect.}

\item{n_hits}{Optional integer. Maximum number of hits to be retained in
summary or exported. Hits are ranked by ascending adjusted p-value, and the
top \code{n_hits} kept.}

\item{hit_criteria}{List of criteria for selecting hits.

Default criteria (if \code{p_adj_NPARC} column is present):\tabular{lrr}{
    \tab  \tab  \cr
    \tab NPARC_pvalue_threshold \tab 0.05 \cr
    \tab DTm_same_sign \tab TRUE \cr
    \tab DTm_gt_Dcontrol \tab TRUE \cr
}


Otherwise:\tabular{lrr}{
    \tab  \tab  \cr
    \tab pvalue_min_threshold \tab 0.05 \cr
    \tab pvalue_max_threshold \tab 0.10 \cr
    \tab DTm_same_sign \tab TRUE \cr
    \tab DTm_gt_Dcontrol \tab TRUE \cr
    \tab slope_threshold \tab -0.6 \cr
}

\itemize{
\item \emph{pvalue_min_threshold:} Threshold \emph{at least one} adjusted
p-value must meet per protein
\item \emph{pvalue_max_threshold:} Threshold \emph{every} adjusted p-value must
meet per protein
\item \emph{NPARC_pvalue_threshold:} Threshold every NPARC-derived p-value must
meet - one pvalue should be generated per protein and condition.
\item \emph{DTm_same_sign:} \eqn{\Delta T_m} must have the same sign
\item \emph{DTm_gt_Dcontrol:} \eqn{\Delta T_m} must be
\eqn{> \Delta T_m (controls)}
\item \emph{slope_threshold:} Steepness threshold for comparison steepest slope.
}}

\item{control_name}{Character. Character string that matches the control
experiment in \code{TPP_tbl$Condition}}

\item{to_export}{Character. File path to save hit data. If \code{NULL},
this will not be saved.}

\item{to_plot}{Boolean. If true, will generate melting point curves of hits
and \code{plot()}}

\item{to_save}{Character. If supplied, will generate melting point curves of
hits and save with \code{\link[=ggsave]{ggsave()}}}

\item{plot_separately}{Boolean. If TRUE, and either \code{to_plot} is TRUE or
\code{to_save} is not \code{NULL}, melting point plots will be generated per protein.
If false, this will be a plot with proteins as facets by default.}

\item{silent}{Boolean. If \code{TRUE}, no console output is shown}

\item{...}{Additional arguments to be passed to \code{\link[=export_TPP]{export_TPP()}},
\code{\link[=predict_melt_curve]{predict_melt_curve()}} or \code{\link[=build_melt_plot_with_model]{build_melt_plot_with_model()}}}
}
\value{
A \code{tibble} containing columns for adjusted pvalue, treated and
control mean \eqn{T_m} and \eqn{\Delta T_m}. One row per protein/condition
combination.
}
\description{
\code{get_TPP_hits()} filters TPP-TR result data to candidates for thermally
stabilised or destabilised proteins, and returns a per-protein summary of
results. Default hit criteria are adapted from those described by Savitsky
et al. 2014 to include the use of \emph{NPARC} p-values - \emph{per} Childs
et al. 2019 - and allow curves that satisfy the non-parametric statistical
tests but don't fit the steep slope pattern required by Savitsky:
\itemize{
\item The FDR-adjusted p-value derived from nonparametric analysis of response
curves \emph{(NPARC)} is less than 0.05
\item All replicates show a positive \eqn{\Delta T_m} or all negative.
\item All \eqn{\Delta T_m} values compared against controls are greater than the
maximum \eqn{\Delta T_m} between controls.
\item No slope threshold is used.
}

Original criteria (that is used if no \emph{NPARC}-derived p-value is
avaiable):
\itemize{
\item The melting point difference (\eqn{\Delta T_m}) compared against control
for a protein has a minimum  \emph{FDR}-adjusted p-value of less than 0.05
and a maximum of less than 0.10 across biological replicates.
\item All replicates show a positive \eqn{\Delta T_m} or all negative.
\item All \eqn{\Delta T_m} values compared against controls are greater than the
maximum \eqn{\Delta T_m} between controls.
\item For each comparison \emph{versus} control, the steepest slope is steeper
than -0.06.
}

Melting point curves for hits will be generated if \code{to_plot} is \code{TRUE}
or \code{to_save} is not \code{NULL}.
}
\examples{
# Minimal data - analysed single-protein dataset
x <- MP_stat_data_ATIC
hit_results_file <- tempfile(pattern = "hits_", fileext = ".xlsx")

# Default criteria return no hits - p-values do not meet criteria
hits <- get_TPP_hits(x, to_export = NULL, to_plot = FALSE)
hits

# Using only subset of hit criteria returns single protein to export as .xlsx
hits <-
  get_TPP_hits(
    x,
    hit_criteria = list(DTm_same_sign = TRUE,
                        DTm_gt_Dcontrol = TRUE),
    to_export = hit_results_file,
    to_plot = FALSE)
hits

unlink(hit_results_file)
}
\references{
Savitski M. M. \emph{et al.}, Tracking cancer drugs in living cells by
thermal profiling of the proteome. \emph{Science}, 346: 1255784 (2014)

Childs, D., \emph{et al.} Non-Parametric Analysis of Thermal Proteome
Profiles Reveals Novel Drug-Binding Proteins. \emph{Molecular & Cellular
Proteomics}, 18, 2506-2515, (2019)
}
