% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_custom_format.R
\name{import_custom_format}
\alias{import_custom_format}
\title{Import TPP data from proteomics tabular data into tidier tibble format.}
\usage{
import_custom_format(
  datafile,
  config,
  path = NULL,
  table_format = c("wide", "long"),
  protein_id_col_name = NULL,
  quantity_pattern = NULL,
  experiment_col_name = NULL,
  quantity_col_name = NULL,
  experiment_id_func = NULL,
  seq_col_name = NULL,
  seq_col_func = NULL,
  silent = FALSE,
  ...
)
}
\arguments{
\item{datafile}{Character. Specify a filename for a delimited text file or
\emph{.xlsx} spreadsheet containing the protein quantity data.}

\item{config}{Either a character, specifying the file name of a delimited
file with experiment configuration, to be opened with
\link{read.table}/\link[readxl:read_excel]{readxl::read_excel} or a \code{data.frame}/\code{tibble} with
configuration information.

This should take the form of a table with the column names
\emph{Experiment}, \emph{Condition}, \emph{Replicate} and
\emph{Temp}.}

\item{path}{Character. Optionally, the path to a directory containing the
data and configuration files. If given, this will be appended to both paths.}

\item{table_format}{One of "wide" or "long". Are the quantity data in the
file to import presented in a wide format - one column per temperature point
observation, or long - one quantity column, one experimental identifier
column, \emph{e.g.}

\emph{Wide:}\tabular{rrr}{
   Quantity_Sample_01.raw \tab Quantity_Sample_02.raw \tab Quantity_Sample_03.raw \cr
   5434942 \tab 5394587 \tab 4523683 \cr
}


\emph{Long:}\tabular{rr}{
   Experiment \tab Quantity \cr
   Sample_01.raw \tab 5434942 \cr
   Sample_02.raw \tab 5394587 \cr
   Sample_03.raw \tab 4523683 \cr
}}

\item{protein_id_col_name}{Character. The exact name of a column in the data
from which a unique proteinn ID will be taken.}

\item{quantity_pattern}{Character. Regular expression that identifies protein
quantity column(s)}

\item{experiment_col_name}{Character. If \code{table_format = "wide"}, the exact
name of a column in the data from which experiment ID should be taken.}

\item{quantity_col_name}{Character. If \code{table_format = "wide"}, the name of
a column in the data from which protein quantity values should be taken.
Will override \code{quantity_pattern} if both are supplied}

\item{experiment_id_func}{Optional function to transform vector of column
names for quantity columns into experiment IDs matching the contents of the
\code{Experiment} column from the \code{config} data. If not given, samples are
assigned a number based on order of appearance in the data.}

\item{seq_col_name}{Optional character. The exact name of a column in the
data from which peptide sequences should be taken. If not given, number of
peptide IDs per protein will be estimated from number of identical rows in
data per protein.}

\item{seq_col_func}{Optional function to transform peptide sequence column to
extract peptide canonical sequence (used to differentiate peptide matches
per protein from total spectrum matches)}

\item{silent}{Boolean. If true, will run silently, without console output.}

\item{...}{Arguments to pass to \code{read.table()} or \code{read_excel()}}
}
\value{
A tibble giving the TPP protein quantity data relative to the lowest
temperature, \eqn{T_1}, as \emph{rel_quantity}, in a tidier format: 1 row
per protein observation.
}
\description{
Import TPP data from proteomics tabular data into tidier tibble format.
}
