% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_spectronaut.R
\name{import_spectronaut}
\alias{import_spectronaut}
\title{Import TPP data from Spectronaut report into tidier tibble format.}
\usage{
import_spectronaut(datafile, config, path = NULL, silent = FALSE)
}
\arguments{
\item{datafile}{Character. Specify a Spectronaut .csv report filename
containing the protein quantity data.}

\item{config}{Either a character, specifying the file name of a text .csv
file with experiment configuration, or literal data that can be passed to
\link[readr:read_delim]{readr::read_csv}. This should take the form of a table with the
column names \emph{Experiment}, \emph{Condition}, \emph{Replicate} and
\emph{Temp}.}

\item{path}{Character. Optionally, the path to a directory containing the
data and configuration files. If given, this will be appended to both paths.}

\item{silent}{Boolean. If true, will run silently, without console output.}
}
\value{
A tibble giving the TPP protein quantity data relative to the lowest
temperature, \eqn{T_1}, as \emph{rel_quantity}, in a tidier format: 1 row
per protein observation.
}
\description{
Import TPP data from Spectronaut report into tidier tibble format.
}
