% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_TPP.R
\name{import_TPP}
\alias{import_TPP}
\title{Import TPP data from chosen proteomics tabular data format into tidier tibble
format}
\usage{
import_TPP(
  datafile,
  config,
  path = NULL,
  import_format = "Spectronaut",
  silent = FALSE,
  ...
)
}
\arguments{
\item{datafile}{Character. Specify a filename for a delimited text file or
\emph{.xlsx} spreadsheet containing the protein quantity data.}

\item{config}{Either a character, specifying the file name of a delimited
file with experiment configuration, to be opened with
\link{read.table} / \link[readxl:read_excel]{readxl::read_excel} or a \code{data.frame} / \code{tibble} with
configuration information.

This should take the form of a table with the column names
\emph{Experiment}, \emph{Condition}, \emph{Replicate} and
\emph{Temp}.}

\item{path}{Character. Optionally, the path to a directory containing the
data and configuration files. If given, this will be appended to both paths.}

\item{import_format}{Character. Specify a format to import with an
\emph{import_} function, or custom specification:
\itemize{
\item "Spectronaut" or "SN" - peptide report from \emph{Spectronaut} DIA protein
quantification (\emph{.csv}): \link{import_spectronaut}
\item "ProteomeDiscover" or "PD" - exported proteins table from
\emph{Thermo Proteome Discoverer} DDA quantification (\emph{.txt}):
\link{import_proteomediscoverer}
\item "Custom" - custom specification: \link{import_custom_format}
}}

\item{silent}{Boolean. If true, will run silently, without console output.}

\item{...}{Arguments to be passed to \code{import_} functions, and then to
\code{read.table()} or \code{read_excel()}}
}
\value{
A tibble giving the TPP protein quantity data relative to the lowest
temperature, \eqn{T_1}, as \emph{rel_quantity}, in a tidier format: 1 row
per protein observation.
}
\description{
\code{import_TPP()} imports tabular proteomics data from a delimited
text file or \emph{.xlsx} spreadsheet and prepares it for TPP data analysis,
by choosing between functions that transform specific formats.
\itemize{
\item Columns specified using the given arguments are selected from the data,
according to the vendor file format
\item Experiment condition combinations (treatment, temperature, \emph{etc.}) are
loaded from a separate config file and matched to recorded data
\item The necessary transformations are performed so the data can be analysed for
\emph{thermal protein profiling (TPP)} data: combining data points so that
a single data point is presented for each experiment combination and
recording the number of peptide hits for each protein as far as that's
possible from the data - this allows the use of minimum matches as a
quality control.
\item The data is reshaped into a tidier long format \code{tibble}, with one
observation per row, one column per statistic.
}
}
\examples{
# Four-protein example - Spectronaut format
sp_report_file <-
  system.file("extdata", "4_protein_peptide_report.csv", package = "tidyTPP")
config_file <-
  system.file("extdata", "4_protein_config.csv", package = "tidyTPP")

# Import data
sp_tbl <-
  import_TPP(sp_report_file, config_file, import_format = "Spectronaut")

sp_tbl
}
