% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MP_data_ATIC}
\alias{MP_data_ATIC}
\title{Parameters for ATIC Protein Melting Curve}
\format{
A data frame with 4 rows and 10 variables:
\describe{
\item{Protein_ID}{Gene name identifying the protein}
\item{Condition}{Treatment given to this sample set}
\item{Replicate}{Replicate number for this sample set}
\item{\code{a}}{parameter \code{a} from sigmoidal curve fitting}
\item{\code{b}}{parameter \code{b} from sigmoidal curve fitting}
\item{melt_point}{melting point, calculated as the model temperature
resulting in 50 \% of the protein destabilised}
\item{infl_point}{inflection point, calculated as the model temperature
for which \eqn{f''(T_{infl}) = 0 }}
\item{slope}{sigmoidal curve slope, calculated as the model temperature
for which \eqn{f'(T_{infl}) }}
\item{plateau}{The plateau of the modeled sigmoidal curve}
\item{R_sq}{\eqn{R^2} for the fitted sigmoidal curve}
}
}
\usage{
MP_data_ATIC
}
\description{
A dataset containing the parameters generated from fitting a melting curve
to raw  proteomics data (DIA) for the protein ATIC:
}
\keyword{datasets}
