% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise_TPP.R
\name{normalise_TPP}
\alias{normalise_TPP}
\title{Normalise TPP-TR Data}
\usage{
normalise_TPP(
  TPP_tbl,
  to_plot = FALSE,
  to_save = NULL,
  quality_filter = NULL,
  norm_subset_filter = NULL,
  quantity_column = "rel_quantity",
  silent = FALSE
)
}
\arguments{
\item{TPP_tbl}{A data frame (or \link[tibble:tibble]{tibble::tibble}) containing proteomics data from a
thermal protein profiling (TPP) experiment. This must contain the columns:
\itemize{
\item \code{Protein_ID}: Unique protein identity
\item \code{Condition} : Category of treated or control sample
\item \code{Replicate} : Replicate number.
\item \code{Temperature}: Experiment temperature in \eqn{\degree C}
\item Protein quantity measurement corresponding to the non-denatured fraction
}

If replicates are not given, a single replicate per condition will be
assumed.}

\item{to_plot}{Boolean. If true, run \code{\link[=plot]{plot()}} on \code{ggplot} object}

\item{to_save}{Character. If supplied, save plot with \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}}}

\item{quality_filter}{A data.frame containing the information needed to build
an initial quantity filter. Columns must be \emph{col}, \emph{lower} and
\emph{upper}. Default:\tabular{lrr}{
   col \tab lower \tab upper \cr
   Pep_N \tab 2 \tab Inf \cr
}


Resulting in filtering for number of detected peptides per protein
\emph{Pep_N} \eqn{\geq 2}}

\item{norm_subset_filter}{A data.frame containing the information needed to
build the protein filter for normP. Columns must be \emph{Temp}, \emph{lower}
and \emph{upper}, \emph{e.g}:\tabular{rrr}{
   Temp \tab lower \tab upper \cr
   56 \tab 0.4 \tab 0.6 \cr
   63 \tab -Inf \tab 0.3 \cr
   67 \tab -Inf \tab 0.2 \cr
}


If no table is given, one will be generated based on the one above, using
the closest temperature points to 56, 63, and 67 \eqn{\degree C},
respectively}

\item{quantity_column}{Character. Name for the column containing
protein quantity values in \code{TPP_tbl}}

\item{silent}{Boolean. If \code{TRUE}, no console output is shown}
}
\value{
A \code{tibble}, as given in TPP_tbl, with the quantity values
(from \code{quantity_column}) replaced with normalised quantities, and an
additional \code{norm_coefficients} column
}
\description{
\code{normalise_TPP()} transforms \emph{Thermal Protein Profiling} (TPP) relative
intensity data, normalising against fitted median melting curves, as
described by Savitsky \emph{et al.} 2014:
\itemize{
\item The set of proteins observed in all experimental groups, \emph{jointP}, is
identified.
\item A subset of proteins is identified, \emph{normP}, with a clear \eqn{T_m}
~ \eqn{56 \degree C} for each experimental group, and the largest
\emph{normP} used.
\item Median fold-change values over all proteins in \emph{normP} are calculated
for each experiment, at each temperature point.
\item Melting curves are fitted, and the curve with the best \eqn{R^2} used as
the normal model.
\item A correction factor is calculated for each median fold-change in order to
coincide with the best-fitting curve.
\item These correction factors are applied to all the protein measurements in the
respective experimental groups.
}
}
\examples{
# Minimal data - four-protein melt curve
x <- quan_data_4prot

# Normalise data
normalise_TPP(x)

# Plot normalisation curves
normalise_TPP(x, to_plot = TRUE)

# Filter proteins to >= 3 observed peptides
normalise_TPP(
  x,
  quality_filter = data.frame(col = "Pep_N", lower = 3, upper = Inf)
)

}
\references{
Savitski M. M. \emph{et al.}, Tracking cancer drugs in living cells by
thermal profiling of the proteome. \emph{Science}, 346: 1255784 (2014)
}
