% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_melt.R
\name{plot_melt}
\alias{plot_melt}
\title{Plot Melt Curve}
\usage{
plot_melt(
  data,
  fit_parameters = NULL,
  to_plot = TRUE,
  to_save = NULL,
  file_suffix_column = NULL,
  to_add_to_ggplot = NULL,
  ...
)
}
\arguments{
\item{data}{TPP protein observation data; see \code{\link[=build_melt_plot_with_model]{build_melt_plot_with_model()}}.
If fit_parameters are not given separately they will be expected as part of
this dataset, as \code{a}, \code{b}, \code{plateau} and, if intending to annotate,
\code{melt_point}, and \code{R_sq} columns}

\item{fit_parameters}{parameters to predict protein melting curve; see
\code{\link[=predict_melt_curve]{predict_melt_curve()}}. Overrides columns in \code{data}}

\item{to_plot}{Boolean. If true, run \code{\link[=plot]{plot()}} on \code{ggplot} object}

\item{to_save}{Character. If supplied, save plot with \code{\link[=ggsave]{ggsave()}}}

\item{file_suffix_column}{Character. The name of a column to be used as an
additional file suffix before the extension if \code{to_save} is given and the
plot is to be saved. The first element of the given column of \code{data} will be
used to identify the plot.}

\item{to_add_to_ggplot}{List. Optional list of further additions to \code{ggplot};
will be added using \code{\link[=+.gg]{+.gg()}}}

\item{...}{Additional arguments to be passed to \code{\link[=predict_melt_curve]{predict_melt_curve()}} or
\code{\link[=build_melt_plot_with_model]{build_melt_plot_with_model()}}}
}
\value{
A data frame or \link{tibble} given in \code{data}
}
\description{
\code{plot_melt} builds and plots a protein melting point from proteomics
\emph{Thermal Protein Profiling} (TPP) data and corresponding fitted model.
Generates prediction data with \code{\link[=predict_melt_curve]{predict_melt_curve()}}, and plot with
\code{\link[=build_melt_plot_with_model]{build_melt_plot_with_model()}} Included options are to plot in an
interactive session, or save using \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}}, while returning the
original data set, to match the behaviour of \emph{e.g.} \code{\link[=print]{print()}} or
\code{\link[=plot]{plot()}}.
}
\examples{
# Minimal data - four-protein melt curve (Protein A only)
x <- quan_data_4prot
norm_x <- normalise_TPP(x)
norm_x <- norm_x[norm_x$Protein_ID == "Protein_A",]
x_parameters <-
analysis_data_4prot[analysis_data_4prot$Protein_ID == "Protein_A",]

# Plot melting point curve
plot_melt(norm_x, x_parameters)

# Use combined data table
x_plus_params <- merge(norm_x, x_parameters)
plot_melt(x_plus_params)

# Customise plot
plot_melt(
  norm_x,
  x_parameters,
  to_add_to_ggplot =
    ggplot2::ggtitle("Protein A",
                     subtitle = "Protein Melting Curve"),
  n_predict = 200,
  annotate = "both",
  facets = FALSE)

}
