% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{quan_data_20prot}
\alias{quan_data_20prot}
\title{Quantity data for 20 Protein Melting Curves}
\format{
A data frame with 800 rows and 8 variables:
\describe{
\item{Protein_ID}{Gene name identifying the protein}
\item{Pep_N}{Number of unique peptides identified for protein}
\item{Match_N}{Number of unique peptide spectral matches (PSMs)
found for protein}
\item{Condition}{Treatment given to this sample set}
\item{Replicate}{Replicate number for this sample set}
\item{Temp}{Experiment temperature in \eqn{\degree C}}
\item{rel_quantity}{Quantity of protein measured, relative to \eqn{T_1}}
\item{raw_quantity}{Raw quantity of protein measured}
}
}
\usage{
quan_data_20prot
}
\description{
A dataset containing the raw proteomics data (DIA) for 20 proteins
}
\keyword{datasets}
