% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{quan_data_ATIC}
\alias{quan_data_ATIC}
\title{Quantity data for ATIC Protein Melting Curve}
\format{
A data frame with 10 rows and 5 variables:
\describe{
\item{Protein_ID}{Gene name identifying the protein}
\item{Condition}{Treatment given to this sample set}
\item{Replicate}{Replicate number for this sample set}
\item{Temp}{Experiment temperature in \eqn{\degree C}}
\item{rel_quantity}{Quantity of protein measured, relative to \eqn{T_1}}
}
}
\usage{
quan_data_ATIC
}
\description{
A dataset containing the parameters generated from fitting a melting curve
to raw  proteomics data (DIA) for the protein ATIC:
}
\keyword{datasets}
