% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_TPP_pipeline.R
\name{apply_TPP_pipeline}
\alias{apply_TPP_pipeline}
\title{Apply tidyTPP TPP-TR pipeline with default settings}
\usage{
apply_TPP_pipeline(
  datafile,
  config,
  path = NULL,
  to_plot = FALSE,
  max_cores = 4
)
}
\arguments{
\item{datafile}{Character. Specify a filename for a delimited text file or
\emph{.xlsx} spreadsheet containing the protein quantity data.}

\item{config}{Either a character, specifying the file name of a delimited
file with experiment configuration, to be opened with
\link{read.table}/\link[readxl:read_excel]{readxl::read_excel} or a \code{data.frame}/\code{tibble} with
configuration information.

This should take the form of a table with the column names
\emph{Experiment}, \emph{Condition}, \emph{Replicate} and
\emph{Temp}.}

\item{path}{Character. Optionally, the path to a directory containing the
data and configuration files. If given, this will be appended to both paths.}

\item{to_plot}{Boolean. Whether to generate plots to the console:
\itemize{
\item Normalisation curves
\item NPARC analysis F-score and p-value distribution
\item Hit protein melting curves
}}

\item{max_cores}{Integer. Maximum number of cores to run analysis using. If
1, will use no parallel or multi-core functions.}
}
\value{
A \code{tibble} of hit proteins, sorted by NPARC p-value
}
\description{
Apply tidyTPP TPP-TR pipeline with default settings: by default imports
Spectronaut report \emph{.csv} file.
}
\examples{
 # Minimal Example - ATIC and CFL1 data
 normP_report_file <-
   system.file("extdata", "normP_report.csv", package = "tidyTPP")
 config_file <-
   system.file("extdata", "ATIC_config.csv", package = "tidyTPP")

 # Analyse using single core, generate and show plots
 normP_hits <-
   apply_TPP_pipeline(datafile = normP_report_file,
                      config = config_file,
                      to_plot = TRUE,
                      max_cores = 1)

 normP_hits
}
