% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_melting_curve.R
\name{fit_melting_curve}
\alias{fit_melting_curve}
\title{Fit Sigmoidal Melting Curve and Find Parameters}
\usage{
fit_melting_curve(
  data,
  x_column = "Temp",
  y_column = "rel_quantity",
  protein_num = NULL,
  protein_total = NULL,
  silent = FALSE,
  show_errors = FALSE,
  curve_fit_method = c("nls", "nls.multstart"),
  ...
)
}
\arguments{
\item{data}{A data frame (or \link[tibble:tibble]{tibble::tibble}) containing proteomics data from a
thermal protein profiling (TPP) experiment. This must contain the columns:
\itemize{
\item Experiment temperature in \eqn{\degree C}
\item Protein quantity measurement corresponding to the non-denatured fraction
}}

\item{x_column}{Character. Name for the column containing x-axis
(temperature) values in \code{data}}

\item{y_column}{Character. Name for the column containing y-axis
(protein quantity) values in \code{data}}

\item{protein_num, protein_total}{Numeric values to be displayed as part of
the progress bar if necessary, ignored if silent = \code{TRUE}

\emph{e.g.} for \code{num} = 30, \code{total} = 256:

\verb{> |=         | 30 of 256}}

\item{silent}{Boolean. If \code{TRUE}, no console output is shown}

\item{show_errors}{Boolean. If \code{TRUE}, errors in calling \code{\link[nls.multstart:nls_multstart]{nls.multstart::nls_multstart()}}
will be shown in console output}

\item{curve_fit_method}{Character. Method for sigmoid curve fitting -
either \emph{"nls"} for a repeated \code{\link[=nls]{nls()}} approach (default) or
\emph{"nls.multstart"} for \code{\link[nls.multstart:nls_multstart]{nls.multstart::nls_multstart()}}.}

\item{...}{Further arguments to be passed to \code{\link[=nls]{nls()}} or \code{\link[nls.multstart:nls_multstart]{nls.multstart::nls_multstart()}}}
}
\value{
A \code{tibble} with 1 row and 13 variables:
\itemize{
\item ID columns from \code{experiment_cols}
\item \code{a}: Parameter \code{a} from fitted sigmoidal curve.
\item \code{b}: Parameter \code{b} from fitted sigmoidal curve.
\item \code{plateau}: Plateau from fitted sigmoidal curve.
\item \code{melt_point}:  Melting point (\eqn{T_m}), temperature at which 50 \% of the
protein is destabilised
\item \code{infl_point}: Inflection point (\eqn{T_{infl}}), temperature at which the
second derivative \eqn{f''(T) = 0}
\item \code{slope}: Sigmoidal curve slope, \emph{i.e.} \eqn{f'(T_{infl})}
\item \code{R_sq}: \eqn{R^2} for the fitted sigmoidal curve
\item \code{RSS}: Residual sum of squares (\emph{RSS})
\item \code{sigma}: Residual standard deviation (\eqn{\sigma})
\item \code{n_coeffs}: Number of coefficients fitted
\item \code{n_obs}: Number of observations in curve
\item \code{log_lik}: log-likelihood (\eqn{\ell})
\item \code{AICc}: Corrected Akaike information criterion (\emph{AICc})
}
}
\description{
\code{fit_melting_curve()} fits observed protein quantity data to a sigmoidal
melting point curve, given by:
\deqn{  f(T)=\frac{1-plateau}{1+e^{-(\frac{a}{T}-b)}}+plateau }

Curve fitting iterates calls to \link{nls} until the starting parameters result in
convergence to a least-squares estimate for best fit.
Arguments given to \code{nls()} by default are:
\itemize{
\item \code{start = c(pl = 0, a = 550, b = 10)}
\item \code{lower = c(pl = 0, a = 0.00001, b = 0.00001)}
\item \code{upper = c(pl = 1.5, a = 15000, b = 250)}
\item \code{na.action = na.omit}
\item \code{algorithm = "port"}
\item \verb{control = stats::nls.control(maxiter=50))}
}

Alternatively, \link[nls.multstart:nls_multstart]{nls.multstart::nls_multstart} can be used to estimate a series of starting
parameters for the non-linear model using a gridstart approach.
Arguments given to \code{nls_multstart()} by default are:
\itemize{
\item \code{iter = c(5,5,5)}
\item \code{start_lower = c(pl = -0.5, a = 0.00001, b = 0.00001)}
\item \code{start_upper = c(pl = 1.5, a = 5000, b = 250)}
\item \code{lower = c(pl = -0.5, a = 0.00001, b = 0.00001)}
}

\eqn{R^2} is calculated using a pseudo-\eqn{R^2} approach for the non-linear
model, \eqn{T_m} from solving \eqn{f(T) = 0.5}, \eqn{T_{infl}} from solving
\eqn{f''(T) = 0}, and slope by evaluating \eqn{f'(T_{infl})}.
}
\examples{
# Minimal data - 4-protein melt curve (Protein A only)
x <- quan_data_4prot[quan_data_4prot$Protein_ID == "Protein_A",]

# Separate condition and replicate
x_sep <- split(x, ~ Condition + Replicate)

# Fit melting curves
total = length(x_sep)
params <-
  lapply(1:total,
    function(i) {
      fit_melting_curve(x_sep[[i]], protein_num = i, protein_total = total)
    }) |>
  Reduce(rbind, x = _)

params

}
\references{
Schellman J. A., The thermodynamics of solvent exchange
\emph{Biopolymers} 34, 1015-1026 (1994)

Savitski M. M. \emph{et al.}, Tracking cancer drugs in living cells by
thermal profiling of the proteome. \emph{Science}, 346: 1255784 (2014)
}
