% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_spectronaut.R
\name{import_spectronaut}
\alias{import_spectronaut}
\title{Import TPP data from Spectronaut report into tidier tibble format.}
\usage{
import_spectronaut(datafile, config, path = NULL, silent = FALSE, ...)
}
\arguments{
\item{datafile}{Character. Specify a Spectronaut .csv report filename
containing the protein quantity data.}

\item{config}{Either a character, specifying the file name of a delimited
file with experiment configuration, to be opened with
\link{read.table}/\link[readxl:read_excel]{readxl::read_excel} or a \code{data.frame}/\code{tibble} with
configuration information.

This should take the form of a table with the column names
\emph{Experiment}, \emph{Condition}, \emph{Replicate} and
\emph{Temp}.}

\item{path}{Character. Optionally, the path to a directory containing the
data and configuration files. If given, this will be appended to both paths.}

\item{silent}{Boolean. If true, will run silently, without console output.}

\item{...}{Arguments to pass to \code{read.table()} or \code{read_excel()}}
}
\value{
A tibble giving the TPP protein quantity data relative to the lowest
temperature, \eqn{T_1}, as \emph{rel_quantity}, in a tidier format: 1 row
per protein observation.
}
\description{
Import TPP data from Spectronaut report into tidier tibble format.
}
\examples{
# Four-protein example - Spectronaut format
sp_report_file <-
  system.file("extdata", "4_protein_peptide_report.csv", package = "tidyTPP")
config_file <-
  system.file("extdata", "4_protein_config.csv", package = "tidyTPP")

# Import data
sp_tbl <- import_spectronaut(sp_report_file, config_file)

sp_tbl
}
