% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_melt.R
\name{plot_melt}
\alias{plot_melt}
\title{Plot Melt Curve}
\usage{
plot_melt(
  data,
  fit_parameters,
  to_plot = TRUE,
  to_save = NULL,
  to_add_to_ggplot = NULL,
  ...
)
}
\arguments{
\item{data}{TPP protein observation data; see \code{\link[=build_melt_plot_with_model]{build_melt_plot_with_model()}}}

\item{fit_parameters}{parameters to predict protein melting curve; see
\code{\link[=predict_melt_curve]{predict_melt_curve()}}}

\item{to_plot}{Boolean. If true, run \code{\link[=plot]{plot()}} on \code{ggplot} object}

\item{to_save}{Character. If supplied, save plot with \code{\link[=ggsave]{ggsave()}}}

\item{to_add_to_ggplot}{List. Optional list of further additions to \code{ggplot};
will be added using \code{\link[=+.gg]{+.gg()}}}

\item{...}{Additional arguments to be parsed to \code{\link[=predict_melt_curve]{predict_melt_curve()}} or
\code{\link[=build_melt_plot_with_model]{build_melt_plot_with_model()}}}
}
\value{
A data frame or \link{tibble} given in \code{data}
}
\description{
\code{plot_melt} builds and plots a protein melting point from proteomics
\emph{Thermal Protein Profiling} (TPP) data and corresponding fitted model.
Generates prediction data with \code{\link[=predict_melt_curve]{predict_melt_curve()}}, and plot with
\code{\link[=build_melt_plot_with_model]{build_melt_plot_with_model()}} Included options are to plot in an
interactive session, or save using \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}}, while returning the
original data set, to match the behaviour of \emph{e.g.} \code{\link[=print]{print()}} or
\code{\link[=plot]{plot()}}.
}
\examples{
# Minimal data - ATIC melt curve
x <- quan_data_ATIC
x_parameters <- MP_data_ATIC

# Plot melting point curve
plot_melt(x, x_parameters)

# Customise plot
plot_melt(
  x,
  x_parameters,
  to_add_to_ggplot = ggplot2::ggtitle("ATIC", subtitle = "Protein Melting Curve"),
  n_predict = 200,
  annotate = "both",
  facets = FALSE)

}
